Parse-Then-Place/requirements.txt
adapter_transformers==3.0.1
deepspeed==0.5.10
lark
matplotlib
numpy
pytorch_fid
PyYAML
scipy
tqdm
transformers==4.19.4
SentencePiece
wandb==0.12.10